backend/requirements.txt
Flask==2.3.3
flask-restx==1.1.0
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
python-dotenv==1.0.0
PyJWT==2.8.0
email-validator==2.1.1
pymongo==4.7.1
Werkzeug==2.3.7

